autoguru-persistence/requirements.txt
flake8
black
isort
mypy
pytest
tortoise-orm